Add a new post transformer method to run after referencesTransformer
The previous solution was in the context_validate, but that is not
executed when running `tofu apply tfplan`. Therefore, during applying an
existing plan with enabled linting was showing nothing.

But the referenceTransformer is executed by all the phases, so having
these linting rules executed at this step allows their run during any
phase.
