+++ internal/tofu/transform_reference.go
@@ -647,6 +647,8 @@ func ReferencesFromConfig(body hcl.Body, schema *configschema.Block) []*addrs.Re
 // PostTransform implements graphNodePostTransform
 func (t *ReferenceTransformer) PostTransform(ctx context.Context, g *Graph) tfdiags.Diagnostics {
 	var diags tfdiags.Diagnostics
+	diags = diags.Append(corelinting.UnusedVariables(ctx, unusedVariables(g)))
+	diags = diags.Append(corelinting.UnusedLocal(ctx, unusedLocals(g)))
 	diags = diags.Append(corelinting.CountInsteadEnabled(ctx, resourceConfigsWithCount(g)))
 	return diags
 }
@@ -669,6 +671,78 @@ func resourceConfigsWithCount(g *Graph) iter.Seq[*configs.Resource] {
 					continue
 				}
 				if !yield(c) {
+					return
+				}
+			}
+		}
+	}
+}
+
+// unusedLocals returns a iter.Seq that will provide all the configs.Local objects for the
+// locals that are detected as being unused.
+// The objects returned are only for the root module.
+// By returning iter.Seq, the analysis is postponed and can be skipped in case the linting rule that
+// needs this data is not enabled by the user.
+func unusedLocals(g *Graph) iter.Seq[*configs.Local] {
+	return func(yield func(*configs.Local) bool) {
+		isUsed := func(n dag.Vertex) bool {
+			for _, u := range g.UpEdges(n) {
+				switch u.(type) {
+				case *nodeCloseModule:
+					// if this is the only reference it has, the vertex is not used
+				default:
+					return true
+				}
+			}
+			return false
+		}
+		for _, v := range g.Vertices() {
+			switch n := v.(type) {
+			case *nodeExpandLocal:
+				if !n.Module.IsRoot() {
+					continue
+				}
+				if isUsed(n) {
+					continue
+				}
+				if !yield(n.Config) {
+					return
+				}
+			}
+		}
+	}
+}
+
+// unusedVariables returns a iter.Seq that will provide all the configs.Variable objects for the
+// variables that are detected as being unused.
+// The objects returned are only for the root module.
+// By returning iter.Seq, the analysis is postponed and can be skipped in case the linting rule that
+// needs this data is not enabled by the user.
+func unusedVariables(g *Graph) iter.Seq[*configs.Variable] {
+	return func(yield func(*configs.Variable) bool) {
+		var isUsed func(n dag.Vertex) bool
+		isUsed = func(n dag.Vertex) bool {
+			for _, u := range g.UpEdges(n) {
+				switch u.(type) {
+				case *nodeCloseModule:
+				// if this is the only reference it has, the vertex is not used
+				case *nodeVariableReference:
+					if isUsed(u) {
+						return true
+					}
+				default:
+					return true
+				}
+			}
+			return false
+		}
+		for _, v := range g.Vertices() {
+			switch n := v.(type) {
+			case *NodeRootVariable:
+				if isUsed(n) {
+					continue
+				}
+				if !yield(n.Config) {
 					return
 				}
 			}
